+++ promgen_k8s/cluster.py
@@ -11,12 +11,10 @@ class Cluster:
       self.bearer_token_file = '/var/run/secrets/kubernetes.io/serviceaccount/token'
       self.ca_file = '/var/run/secrets/kubernetes.io/serviceaccount/ca.crt'
       self.api_server = 'kubernetes.default.svc'
-      # no self.proxy
     else:
       self.bearer_token_file = '/var/run/kube_secrets/{0}_bearer_token'.format(self.name)
       self.ca_file = '/var/run/kube_secrets/{0}_ca_crt'.format(self.name)
       self.api_server = 'api.internal.{0}.{1}'.format(self.name, self.public_domain)
-      self.proxy = 'prometheus-proxy.{0}.{1}'.format(self.name, self.private_domain)
 
   def get_kubernetes_sd_config(self, role):
     if self.incluster: