+++ promgen_k8s/pods_job.py
@@ -27,7 +27,17 @@ class PodsJob:
   def generate_interval(self, prom_conf, c, interval_name, interval_value):
     prom_conf['scrape_configs'].append({
       'job_name': '{0}-kubernetes-pods-{1}'.format(c.name, interval_name),
+      'scheme': 'https',
       'kubernetes_sd_configs': [ c.get_kubernetes_sd_config('pod') ],
+
+      # This TLS & bearer token file config is used to connect to the actual scrape
+      # endpoints for cluster components. This is separate to discovery auth
+      # configuration because discovery & scraping are two separate concerns in
+      # Prometheus. The discovery auth config is automatic if Prometheus runs inside
+      # the cluster. Otherwise, more config options have to be provided within the
+      # <kubernetes_sd_config>.
+      'tls_config': { 'ca_file': c.ca_file },
+      'bearer_token_file': c.bearer_token_file,
 
       'relabel_configs': [
         keep(source_labels=['__meta_kubernetes_pod_annotation_prometheus_io_scrape'], regex=True),
@@ -36,8 +46,13 @@ class PodsJob:
           regex='(.+)',
           target_label='__metrics_path__'),
         replace(source_labels=['__address__', '__meta_kubernetes_pod_annotation_prometheus_io_port'],
-          regex='([^:]+)(?::\\d+)?;(\\d+)', replacement='$1:$2',
+          separator=';', regex='([^:]+)(?::\\d+)?;(\\d+)', replacement='$1:$2',
           target_label='__address__'),
+        replace(source_labels=['__meta_kubernetes_namespace', '__meta_kubernetes_pod_name', '__meta_kubernetes_pod_annotation_prometheus_io_port', '__metrics_path__'],
+          separator=';', regex='(.+);(.+);(.+);(.+)', replacement='/api/v1/namespaces/$1/pods/$2:$3/proxy$4',
+          target_label='__metrics_path__'),
+        copy_value('__address__', 'instance'),
+        set_value('__address__', '{0}:443'.format(c.api_server)),
         labelmap(regex='__meta_kubernetes_pod_label_(.+)'),
         copy_value('__meta_kubernetes_namespace', 'kubernetes_namespace'),
         copy_value('__meta_kubernetes_pod_name', 'kubernetes_pod_name'),
@@ -50,15 +65,6 @@ class PodsJob:
         drop(source_labels=['__name__'], regex='go_.*')
       ]
     })
-
-    if not c.incluster:
-      prom_conf['scrape_configs'][-1]['relabel_configs'].extend([
-        copy_value('__address__', 'instance'),
-        replace(source_labels=['__address__','__metrics_path__'],
-          regex='([0-9]*\\.[0-9]*\\.[0-9]*\\.[0-9]*:[0-9]*);(.*)', replacement='/proxy/$1$2',
-          target_label='__metrics_path__'),
-        set_value('__address__', c.proxy)
-      ])
 
     if interval_name is 'default':
       prom_conf['scrape_configs'][-1]['relabel_configs'][1] = \