+++ promgen_k8s/services_job.py
@@ -17,32 +17,33 @@ class ServicesJob:
   def generate(self, prom_conf, c):
     prom_conf['scrape_configs'].append({
       'job_name': '{0}-kubernetes-services'.format(c.name),
+      'scheme': 'https',
       'kubernetes_sd_configs': [ c.get_kubernetes_sd_config('service') ],
 
-      'metrics_path': '/probe',
+      # This TLS & bearer token file config is used to connect to the actual scrape
+      # endpoints for cluster components. This is separate to discovery auth
+      # configuration because discovery & scraping are two separate concerns in
+      # Prometheus. The discovery auth config is automatic if Prometheus runs inside
+      # the cluster. Otherwise, more config options have to be provided within the
+      # <kubernetes_sd_config>.
+      'tls_config': { 'ca_file': c.ca_file },
+      'bearer_token_file': c.bearer_token_file,
+
+      'metrics_path': '/api/v1/namespaces/monitoring/services/blackbox-exporter/proxy/probe',
       'params': {
         'module': ['http_2xx']
       },
 
       'relabel_configs': [
         keep(source_labels=['__meta_kubernetes_service_annotation_prometheus_io_probe'], regex=True),
-        None,
+        copy_value('__address__', '__param_target'),
         copy_value('__address__', 'instance'),
-        set_value('__address__', 'blackbox-exporter'),
+        set_value('__address__', '{0}:443'.format(c.api_server)),
         labelmap(regex='__meta_kubernetes_service_label_(.+)'),
         copy_value('__meta_kubernetes_namespace', 'kubernetes_namespace'),
         copy_value('__meta_kubernetes_service_name', 'kubernetes_service_name')
       ]
     })
-
-    if c.incluster:
-      prom_conf['scrape_configs'][-1]['relabel_configs'][1] = \
-        copy_value('__address__', '__param_target')
-    else:
-      prom_conf['scrape_configs'][-1]['relabel_configs'][1] = \
-        replace(source_labels=['__address__'],
-          regex='(.*)', replacement=c.proxy+'/proxy/$1/',
-          target_label='__param_target')
 
     # set job's scrape_interval if defined
     if not self.scrape_interval is None: